Implement document detail page (#10)
Fetch document data when the DocumentDetail component is mounted.

Before, the snapshot data only passed when users clicked the link.
Implement to fetch data when the DocumentDetail component is mounted
to show the document data when entering the URL directly.

This commit also adds some codes by pasting the grpc protocol.
(to keep a single source of truth)
- Add project list command
- Remove collection from document
- Rename agent to server 
- Rename `db` package to `database`

Co-authored-by: Youngteac Hong <youngteac.hong@navercorp.com>