Replace CreateReactApp with Vite (#153)
This commit updates libraries and replaces CreateReactApp with Vite.
Currently dashboard uses React 17, and after being in use for
approximately two years, there is a need to upgrade the stack.
CreateReactApp is not the recommended approach by React team for now,
and Vite is considered the most direct alternative to replace it.
Additionally, Jest is replaced by Vitest to align with the technology
upgrade.

---------

Co-authored-by: Yourim Cha <yourim.cha@navercorp.com>